[js-api] Define multiple return value semantics
- When a JS function is imported by Wasm, with a multiple return
  value type, its return value is iterated over to get the multiple
  WebAssembly values.
- When a WebAssembly function is exported to JS, multiple return
  values become an Array of those values.

These semantics were approved by a poll at the 2018 WebAssembly CG
TPAC meeting.

The wording here is based on JS operations rather than WebIDL;
this is because the operations ToJSValue/ToWebAssemblyValue are
based on the JS type Number rather than IDL types like double.
However, the semantics are in parallel with what it means to
convert from and to a WebIDL sequence.