Imported JS calls don't handle differing arities. (#20)
Right now there is a spec assert that host call's iterable result and wasm have the same arity. There's no guarantee this is true. Instead we should probably throw a type error if they don't match. In the future, we could do the more JS-y approach and allow the arity of result to be at least the arity of the expected signature.