Check multiple returns in generated JS code
This change modifies `assert_return`. After the Wasm function is
invoked, the result will either be `undefined`, a JS `Number`, or a JS
`Array`. It's simpler to treat them all as an array, with `undefined`
representing an empty array, and `Number` representing an array with one
element.

We don't check for `Number` here, since as soon as we add support for
the reference-types proposal, a scalar result may be an `Object` too.

Fixes issue #44.